requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
beautifulsoup4==4.8.2
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-celery==3.3.1
django-cleanup==4.0.0
django-optimized-image==0.2.0
django-sass==0.2.0
django-translations==1.1.2
env-tools==2.2.0
gunicorn==20.0.4
idna==2.8
kombu==4.6.7
libsass==0.19.4
Pillow==7.0.0
psycopg2-binary
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
structlog==19.2.0
tinify==1.5.1
urllib3==1.25.7
vine==1.3.0
whitenoise==5.0.1
